lots of experimentation
