+++ src/ewatercycle/plugins/wflow/ewc_model.py
@@ -15,7 +15,7 @@ from grpc4bmi.bmi_client_singularity import BmiClientSingularity
 
 from ewatercycle import CFG
 from ewatercycle.plugins.wflow.forcing import WflowForcing
-from ewatercycle.models.abstract import AbstractModel
+from ewatercycle.models.abstract import DefaultModel
 from ewatercycle.parameter_sets import ParameterSet
 from ewatercycle.parametersetdb.config import CaseConfigParser
 from ewatercycle.util import find_closest_point, get_time, to_absolute_path
@@ -23,7 +23,7 @@ from ewatercycle.util import find_closest_point, get_time, to_absolute_path
 logger = logging.getLogger(__name__)
 
 
-class Wflow(AbstractModel[WflowForcing]):
+class Wflow(DefaultModel[WflowForcing]):
     """Create an instance of the Wflow model class.
 
     Args:
